Stripped-down version with icosoc routines
Version now is below 128k but doesn't include most application-specific
interesting stuff.
The icosoc routines for I/O are now linked in (thanks to Clifford for
spotting this) and we have debug output.
Current code is enough to get cooperative multitasking and timers
working and tested, this is next todo.